- Fixed bug in wind initialisation/calculation - AAT start/finish radials step in 1 degree increments - Fixes for build on VS2005 (PC) - Fixed various out of bound bugs for task waypoints

+++ Common/Source/MapWindow2.cpp
@@ -653,8 +653,8 @@ void MapWindow::CalculateScreenPositions(POINT Orig, RECT rc,
                       DrawInfo.Latitude, 
                       *Orig_Aircraft);
 
-        if ((fabs(Orig_Aircraft->x-screen.x)<(rc.right-rc.left)/3)
-            && (fabs(Orig_Aircraft->y-screen.y)<(rc.bottom-rc.top)/3)) {
+        if ((fabs((double)Orig_Aircraft->x-screen.x)<(rc.right-rc.left)/3)
+            && (fabs((double)Orig_Aircraft->y-screen.y)<(rc.bottom-rc.top)/3)) {
           
         } else {
           // out of bounds, center on aircraft
